requirements.txt
sklearn
imblearn
pandas
matplotlib
seaborn
pyyaml
black
pytest
pre-commit
